requirements.txt
# Flask Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.21
PyMySQL==1.1.0

# Security
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography==41.0.4

# Background Tasks
APScheduler==3.10.4

# HTTP Requests
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Blockchain/Crypto
tronapi==3.1.6

# Utilities
python-jose==3.3.0
Pillow==10.0.1
qrcode==7.4.2

# Development
pytest==7.4.2
pytest-flask==1.2.0

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Timezone
pytz==2023.3

# JSON Web Tokens
PyJWT==2.8.0

# Decimal handling (built-in Python module)
# decimal is built-in, no need to install

# Date utilities
python-dateutil==2.8.2
